minor refactor of helpers
