+++ examples/with-solana/src/utils/createSolanaWallet.ts
