+++ examples/with-solana/src/utils/index.ts
@@ -1,0 +1,14 @@
+import { createNewSolanaWallet } from "./createSolanaWallet";
+import * as solanaNetwork from "./solanaNetwork";
+import { signMessage } from "./signMessage";
+import { createAndSignTransfer, signTransfers } from "./createSolanaTransfer";
+import { print } from "./print";
+
+export {
+  createNewSolanaWallet,
+  solanaNetwork,
+  createAndSignTransfer,
+  signTransfers,
+  signMessage,
+  print,
+};
