+++ examples/with-solana/src/utils/print.ts
