+++ examples/with-solana/src/utils/signMessage.ts
